backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
requests==2.31.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.4.1
pydantic==2.6.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
alembic==1.13.1
